# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: JuneXDeployment API — Discord bot hosting platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /deployments:
    get:
      operationId: listDeployments
      tags: [deployments]
      summary: List current user's deployments
      responses:
        "200":
          description: List of deployments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Deployment"
    post:
      operationId: createDeployment
      tags: [deployments]
      summary: Create and deploy a bot
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DeploymentInput"
      responses:
        "201":
          description: Deployment created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
  /deployments/{id}:
    get:
      operationId: getDeployment
      tags: [deployments]
      summary: Get a deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deployment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteDeployment
      tags: [deployments]
      summary: Delete a deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /deployments/{id}/start:
    post:
      operationId: startDeployment
      tags: [deployments]
      summary: Start a stopped deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
  /deployments/{id}/stop:
    post:
      operationId: stopDeployment
      tags: [deployments]
      summary: Stop a running deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
  /deployments/{id}/restart:
    post:
      operationId: restartDeployment
      tags: [deployments]
      summary: Restart a deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Restarted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
  /deployments/{id}/logs:
    get:
      operationId: getDeploymentLogs
      tags: [deployments]
      summary: Get logs for a deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeploymentLogs"
  /deployments/{id}/env:
    patch:
      operationId: updateDeploymentEnv
      tags: [deployments]
      summary: Update environment variables for a deployment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EnvVarsUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Deployment"
...
  /payments/initiate:
    post:
      operationId: initiatePayment
      tags: [payments]
      summary: Initiate a Paystack card payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentInitiateInput"
      responses:
        "200":
          description: Payment initiated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentInitiateResponse"
  /payments/stk-push:
    post:
      operationId: stkPush
      tags: [payments]
      summary: M-Pesa STK push via Paystack
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StkPushInput"
      responses:
        "200":
          description: STK push sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MpesaPushResult"
  /payments/verify/{reference}:
    get:
      operationId: verifyPayment
      tags: [payments]
      summary: Verify a Paystack payment
      parameters:
        - name: reference
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Payment status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentVerifyResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    Deployment:
      type: object
      required: [id, userId, templateId, templateName, botName, status, createdAt]
      properties:
        id:
          type: integer
        userId:
          type: integer
        templateId:
          type: integer
        templateName:
          type: string
        botName:
          type: string
        herokuAppId:
          type: ["string", "null"]
        status:
          type: string
          enum: [queued, building, online, offline, error, suspended]
        createdAt:
          type: string
        updatedAt:
          type: string
    DeploymentInput:
      type: object
      required: [templateId, botName, envVars]
      properties:
        templateId:
          type: integer
        botName:
          type: string
        envVars:
          type: object
          additionalProperties:
            type: string
    DeploymentLogs:
      type: object
      required: [lines]
      properties:
        lines:
          type: array
          items:
            type: string
    EnvVarsUpdate:
      type: object
      required: [envVars]
      properties:
        envVars:
          type: object
          additionalProperties:
            type: string
...
    PaymentInitiateInput:
      type: object
      required: [email, amount, planType]
      properties:
        email:
          type: string
        amount:
          type: integer
        planType:
          type: string
        currency:
          type: string
        callbackUrl:
          type: string
    PaymentInitiateResponse:
      type: object
      required: [authorizationUrl, reference, accessCode]
      properties:
        authorizationUrl:
          type: string
        reference:
          type: string
        accessCode:
          type: string
    StkPushInput:
      type: object
      required: [phone, amount, email]
      properties:
        phone:
          type: string
        amount:
          type: integer
        email:
          type: string
        planType:
          type: string
    MpesaPushResult:
      type: object
      required: [status, reference]
      properties:
        status:
          type: string
        reference:
          type: string
        message:
          type: string
    PaymentVerifyResponse:
      type: object
      required: [status, reference]
      properties:
        status:
          type: string
        reference:
          type: string
        amount:
          type: integer
        email:
          type: string
        paidAt:
          type: ["string", "null"]